agri_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.6
django-cors-headers==4.8.0
django-ipware==7.0.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
fpdf2==2.8.5
future==1.0.0
geocoder==1.38.1
# google-ai-generativelanguage==0.6.10
# google-api-core==2.28.1
# google-api-python-client==2.187.0
# google-auth==2.45.0
# google-auth-httplib2==0.3.0
# google-generativeai==0.8.3
openai
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
kiwisolver==1.4.9
markdown2==2.5.4
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pillow==12.0.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
ratelim==0.1.6
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.6.0
